+++ crates/protos/src/protos/mod.rs
@@ -746,6 +746,7 @@ pub mod coresdk {
                     Self {
                         status: Some(aer::Status::Failed(Failure {
                             failure: Some(fail),
+                            ..Default::default()
                         })),
                     }
                 }
@@ -830,7 +831,10 @@ pub mod coresdk {
                     Self {
                         status: match r {
                             Ok(p) => Some(aer::Status::Completed(Success { result: Some(p) })),
-                            Err(f) => Some(aer::Status::Failed(Failure { failure: Some(f) })),
+                            Err(f) => Some(aer::Status::Failed(Failure {
+                                failure: Some(f),
+                                ..Default::default()
+                            })),
                         },
                     }
                 }
@@ -866,6 +870,7 @@ pub mod coresdk {
                     match self.status {
                         Some(activity_resolution::Status::Failed(Failure {
                             failure: Some(ref f),
+                            ..
                         })) => f.is_timeout(),
                         _ => None,
                     }